- fix onboarding

+++ frontend/src/app/routes/home/home.component.ts
@@ -88,45 +88,46 @@ export class HomeComponent implements OnInit {
     const hasSeenTour = localStorage.getItem('hasSeenTour');
 
     if (!hasSeenTour) {
+      const stepElements = ['step1', 'step2', 'step3', 'step4', 'step5', 'step6', 'step7', 'step8', 'step9'];
+      const allElementsExist = stepElements.every(id => document.getElementById(id) !== null);
 
-      const intro = introJs();
-      intro.setOptions({
-        nextLabel: 'Siguiente',
-        prevLabel: 'Atrás',
-        doneLabel: 'Finalizar',
-        exitOnEsc: true,
-        exitOnOverlayClick: false
-      });
+      if (allElementsExist) {
+        const intro = introJs();
 
-      intro.start();
+        intro.setOptions({
+          nextLabel: 'Siguiente',
+          prevLabel: 'Atrás',
+          doneLabel: 'Finalizar',
+          exitOnEsc: true,
+          exitOnOverlayClick: false,
+          showProgress: true,
+          steps: [
+            { intro: "👋 ¡Bienvenido al Dashboard de Lytics! Aquí puedes ver toda la información de consumo energético de tu hogar." },
+            { element: '#step1', intro: "🕒 Reloj – Indica el horario actual en Buenos Aires." },
+            { element: '#step2', intro: "💧 Humedad – Muestra el nivel actual de humedad relativa en el ambiente." },
+            { element: '#step3', intro: "🌡️ Temperatura – Indica la temperatura ambiente en grados Celsius." },
+            { element: '#step4', intro: "💵 Consumo Real – Visualiza el consumo eléctrico actual expresado en kilovatios-hora." },
+            { element: '#step5', intro: "PANEL – Información en tiempo real proporcionada por tu dispositivo." },
+            { element: '#step6', intro: "⚡ Voltaje – Visualiza la tensión eléctrica que llega a tus dispositivos." },
+            { element: '#step7', intro: "🔌 Amperaje – Refleja la cantidad de corriente eléctrica fluyendo en el sistema." },
+            { element: '#step8', intro: "💡 Watts – Indica el consumo instantáneo de energía, similar al combustible utilizado en tiempo real." },
+            { element: '#step9', intro: "⚙️ Kilovatios-hora (kWh) – Mide el consumo acumulado de electricidad durante un período específico." },
+            { element: '#step5', intro: "Ahora podés visualizar detalladamente tus consumos de energía en tiempo real." }
+          ]
+        });
 
-      intro.setOptions({
-        steps: [
-          {intro: "👋 ¡Bienvenido al Dashboard de Lytics! Aquí puedes ver toda la información de consumo energético de tu hogar."},
-          {element: '#step1', intro: "🕒 Reloj – Indica el horario actual en Buenos Aires."},
-          {element: '#step2', intro: "💧 Humedad – Muestra el nivel actual de humedad relativa en el ambiente."},
-          {element: '#step3', intro: "🌡️ Temperatura – Indica la temperatura ambiente en grados Celsius."},
-          {element: '#step4', intro: "💵 Consumo Real – Visualiza el consumo eléctrico actual expresado en kilovatios-hora."},
-          {element: '#step5', intro: "PANEL – Información en tiempo real proporcionada por tu dispositivo."},
-          {element: '#step6', intro: "⚡ Voltaje – Visualiza la tensión eléctrica que llega a tus dispositivos."},
-          {element: '#step7', intro: "🔌 Amperaje – Refleja la cantidad de corriente eléctrica fluyendo en el sistema.\n"},
-          {element: '#step8', intro: "💡 Watts – Indica el consumo instantáneo de energía, similar al combustible utilizado en tiempo real."},
-          {element: '#step9', intro: "⚙️ Kilovatios-hora (kWh) – Mide el consumo acumulado de electricidad durante un período específico."},
-          {element: '#step5', intro: "Ahora podés visualizar detalladamente tus consumos de energía en tiempo real."}
-        ],
-        showProgress: true,
-        exitOnOverlayClick: false
-      });
+        intro.start();
 
-      intro.start();
+        intro.oncomplete(() => {
+          localStorage.setItem('hasSeenTour', 'true');
+        });
 
-      intro.oncomplete(() => {
-        localStorage.setItem('hasSeenTour', 'true');
-      });
-
-      intro.onexit(() => {
-        localStorage.setItem('hasSeenTour', 'true');
-      });
+        intro.onexit(() => {
+          localStorage.setItem('hasSeenTour', 'true');
+        });
+      } else {
+        setTimeout(() => this.startTour(), 500);
+      }
     }
   }
 }
